requirements.txt
beautifulsoup4~=4.12.2
pydantic~=2.3.0
pytz~=2021.1
pyTelegramBotAPI~=3.8.3
psycopg2-binary~=2.9.7
requests~=2.26.0
tqdm~=4.66.1
wheel~=0.37.0
wrapt~=1.15.0